fix(nav): cast Redis-cached sidebar counters to int (fixes 500 on every page)
Sidebar composers declared `: int` returns but returned the value straight
from Cache::remember(). Under the Redis cache store, numeric values are stored
as-is and read back as PHP strings (RedisStore (un)serialize skips is_numeric
values), so on any cache hit the composer returned a string and PHP threw
"Return value must be of type int, string returned" while rendering
layouts/app.blade.php — i.e. on every page that shows the sidebar.

Cast the Cache::remember() result to (int) in both SidebarComposer and
SiteSwitcherComposer. Cache-miss already returned int; the cast makes cache-hit
match. switcherSites() (a Collection) is unaffected.

+++ app/View/Composers/SidebarComposer.php
@@ -54,7 +54,7 @@ class SidebarComposer
      */
     private function updatesCount(User $user): int
     {
-        return Cache::remember("sidebar:updates_count:user:{$user->id}", 60, function () use ($user) {
+        return (int) Cache::remember("sidebar:updates_count:user:{$user->id}", 60, function () use ($user) {
             $connectedVisibleSite = fn ($q) => $q
                 ->where('is_connected', true)
                 ->visibleTo($user);
@@ -77,7 +77,7 @@ class SidebarComposer
      */
     private function securityCount(User $user): int
     {
-        return Cache::remember("sidebar:security_count:user:{$user->id}", 60, function () use ($user) {
+        return (int) Cache::remember("sidebar:security_count:user:{$user->id}", 60, function () use ($user) {
             return SecurityIssue::query()
                 ->active()
                 ->severity('critical')
@@ -93,7 +93,7 @@ class SidebarComposer
      */
     private function errorLogsCount(User $user): int
     {
-        return Cache::remember("sidebar:error_logs_count:user:{$user->id}", 60, function () use ($user) {
+        return (int) Cache::remember("sidebar:error_logs_count:user:{$user->id}", 60, function () use ($user) {
             return PhpErrorLog::query()
                 ->unresolved()
                 ->whereHas('site', fn ($q) => $q->visibleTo($user))
